requirements.txt
pandas
plotly
pydantic
langchain-openai
langchain-core
matplotlib
langgraph
python-dotenv
